+++ src/models/user.js
@@ -42,13 +42,40 @@ export const getUserWithPassword = async (email) => {
   return result.rows[0];
 };
 
-// Obtener usuario por email
+// Obtener usuario por email con todos los campos
 export const getUserByEmail = async (email) => {
   const result = await pool.query(
-    'SELECT id_user, name, last_name, email, id_role FROM users WHERE email = $1',
+    'SELECT * FROM users WHERE email = $1',
     [email]
   );
   return result.rows[0]; // Retorna el usuario encontrado o undefined
+};
+
+// Verificar si el email ya existe antes de editar
+export const getUserByEmailEdit = async (email) => {
+  const result = await pool.query('SELECT * FROM users WHERE email = $1', [email]);
+  return result.rows[0];
+};
+
+// Actualizar datos del usuario 
+export const updateUser = async (email, userData) => {
+  // Filtrar campos `undefined` o `null`
+  const validFields = Object.entries(userData).filter(([_, value]) => value !== undefined && value !== null);
+
+  if (validFields.length === 0) {
+    throw new Error('No se enviaron datos válidos para actualizar');
+  }
+
+  // Construcción dinámica de la consulta
+  const fields = validFields.map(([key], index) => `${key} = $${index + 1}`).join(', ');
+  const values = validFields.map(([_, value]) => value);
+
+  const result = await pool.query(
+    `UPDATE users SET ${fields} WHERE email = $${values.length + 1} RETURNING *`,
+    [...values, email]
+  );
+
+  return result.rows[0];
 };
 
 // Actualizar rol de usuario
@@ -58,4 +85,23 @@ export const updateUserRole = async (id_user, newRoleId) => {
     [newRoleId, id_user]
   );
   return result.rows[0]; // Retorna el usuario actualizado o undefined
+};
+
+
+// //buscar usuario por email para eliminarlo
+// export const getUserByEmail = async (email) => {
+//   const result = await pool.query(
+//     'SELECT * FROM users WHERE email = $1',
+//     [email]
+//   );
+//   return result.rows[0];
+// };
+
+//eliminar usuario por email
+export const deleteUserByEmail = async (email) => {
+  const result = await pool.query(
+    'DELETE FROM users WHERE email = $1 RETURNING *',
+    [email]
+  );
+  return result.rows[0]; // Retorna el usuario eliminado
 };
